fixed thread priority
priority worked correctly but the product ID did not print correctly, this was due to the initializing of the product ID occuring in the constructor. Because the priority did not control the order of object creation the ininitialization and incrementing of the product ID had to be moved into the run method.